pack_box: corrected object placement on background
Stopped removing alpha channel when reading images of the objects.
Moved function putObjectToBackground to a separate file and added
handling transparancy there, so that if image of an object has
some transparancy in it, this transparancy is considered when
adding the object to the scene. Added shuffling of object indeces
when adding them to the scene, so that the order of their addition
is not predefined. Added limit to attempts to put object to the
scene without intersections - if it is not possible to put the
object in some attempts, it is skipped.